Final version

+++ Trabalho2.X/Real.c
@@ -1,3 +1,4 @@
+#include <math.h>
 #include <xc.h>
 #include <sys/attribs.h>
 #include <stdio.h>
@@ -10,13 +11,14 @@ float ADC_OUT(void);
 float transf_function(float val);
 
 /* Function ReadSensor is used to obtain the real value of temperature from PT100*/
-float ReadSensor(void){
+double ReadSensor(void){
    // Variable declarations;
     double temperature_real;
     double duty;
     /*dont necessary realized mean because is already done in ADC_OUT()*/
-    start_ADC();                                 // init conversation           
-    duty = ADC_OUT();                            // obtain the value from ADC
-    temperature_real = transf_function(duty);     
+    start_ADC();                                        // init conversation           
+    duty = ADC_OUT();                                   // obtain the value from ADC
+    // temperature_real = transf_function(0.123242);    // test if the temperature is 60ºC, table R is 123.242 ohm and knowing the current (1mA) we get V = 0.123242
+    temperature_real = transf_function(duty);
     return temperature_real;
 }
